backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
psycopg2-binary>=2.9
python-dotenv>=1.0
mercadopago>=2.0
requests>=2.31
gunicorn>=21.0
Pillow>=10.0
